# Use string.Create in PhysicalAddress.ToString