--- a/src/System.Net.NetworkInformation/src/System/Net/NetworkInformation/PhysicalAddress.cs
+++ b/src/System.Net.NetworkInformation/src/System/Net/NetworkInformation/PhysicalAddress.cs
@@ -86,28 +86,16 @@ namespace System.Net.NetworkInformation
 
         public override string ToString()
         {
-            StringBuilder addressString = new StringBuilder();
-
-            foreach (byte value in _address)
+            return string.Create(_address.Length * 2, _address, (span, addr) =>
             {
-                int tmp = (value >> 4) & 0x0F;
-
-                for (int i = 0; i < 2; i++)
+                int p = 0;
+                foreach (byte value in addr)
                 {
-                    if (tmp < 0x0A)
-                    {
-                        addressString.Append((char)(tmp + 0x30));
-                    }
-                    else
-                    {
-                        addressString.Append((char)(tmp + 0x37));
-                    }
-
-                    tmp = ((int)value & 0x0F);
+                    byte upper = (byte)(value >> 4), lower = (byte)(value & 0xF);
+                    span[p++] = (char)(upper + (upper < 10 ? '0' : 'A' - 10));
+                    span[p++] = (char)(lower + (lower < 10 ? '0' : 'A' - 10));
                 }
-            }
-
-            return addressString.ToString();
+            });
         }
 
         public byte[] GetAddressBytes()