Simple change to avoid an unnecessary StringBuilder.

```C#
private readonly PhysicalAddress _short = new PhysicalAddress(new byte[1] { 42 });
private readonly PhysicalAddress _long = new PhysicalAddress(Enumerable.Range(0, 256).Select(i => (byte)i).ToArray());

[Benchmark] public void PAShort() => _short.ToString();
[Benchmark] public void PALong() => _long.ToString();
```

Before:
```
  Method |        Mean |      Error |    StdDev |      Median |  Gen 0 | Allocated |
-------- |------------:|-----------:|----------:|------------:|-------:|----------:|
 PAShort |    36.64 ns |  0.7689 ns |  1.622 ns |    36.00 ns | 0.0324 |     136 B |
  PALong | 2,266.73 ns | 33.7816 ns | 31.599 ns | 2,259.47 ns | 0.5951 |    2504 B |
```

After:
```
  Method |      Mean |      Error |     StdDev |  Gen 0 | Allocated |
-------- |----------:|-----------:|-----------:|-------:|----------:|
 PAShort |  13.30 ns |  0.2307 ns |  0.2158 ns | 0.0076 |      32 B |
  PALong | 679.91 ns | 12.8956 ns | 12.0626 ns | 0.2489 |    1048 B |
```

cc: @davidsh, @rmkerr